manual_transaction: Move the module to server.lib namespace
