+++ server/lib/manual_transactions.py
